requirements.txt
flet
requests
PyGithub
python-dotenv
sentence-transformers
transformers
torch
numpy
psutil
modelscope